Partial Changes

+++ src/Services/Domain.Services/RoomService.js
@@ -2,7 +2,8 @@ import { TypeOfStay } from "../../Constants/Enum/Rooms";
 const evernode = require("evernode-js-client");
 const settings = require("../../settings.json").settings;
 const constants = require("../../Constants/Constants");
-const { SqliteDatabase } = require("../Common.Services/dbHandler");
+const { SqliteDatabase } = require("../Common.Services/dbHandler").default;
+import { SharedService } from "../Common.Services/SharedService";
 
 export class RoomService {
 	#message = null;
@@ -16,7 +17,7 @@ export class RoomService {
 		this.#dbContext = new SqliteDatabase(this.#dbPath);
 	}
 
-	async createUpdateRoomType() {
+	async createRoomType() {
 		let resObj = {};
 		let roomTypeId
 		const rooms = {};
@@ -39,12 +40,14 @@ export class RoomService {
 		// Saving to the roomType table
 		try {
 			roomTypeId = (await this.#dbContext.insertValue(Tables.ROOMTYPES, roomType)).lastId;
+			console.log("roomTypeId at creation:",roomTypeId);
 		} catch (error) {
 			throw new Error("Error occured in room type saving");
 		}
 
 		if (roomTypeId != null) {
-			// Saving to the roomTypemage table
+			console.log("Saving to the roomTypeImage table");
+			// Saving to the roomTypeImage table
 			if (data.ImageUrls && data.ImageUrls.length > 0) {
 				for (const url of data.ImageUrls) {
 					const imageEntity = {
@@ -66,6 +69,7 @@ export class RoomService {
 
 			// Saving pricing details
 			if (data.RoomCodes && data.RoomCodes.length == data.RoomsCount) {
+				console.log("Saving to the Pricing table");
 				for (const code of data.RoomCodes) {
 					rooms = {
 						RoomTypeId: roomTypeId,
@@ -97,8 +101,29 @@ export class RoomService {
 			}
 		}
 
+		this.#dbContext.close();
 		resObj.success = { rowId: roomTypeId };
 		return resObj;
+	}
+	
+	async editRoomType() {
+		let resObj = {};
+		const data = this.#message.data;
+
+		try {
+			this.#dbContext.open();
+			const roomType = await this.#dbContext.findById(Tables.ROOMTYPES, data.roomTypeId);
+			if (!roomType) {
+				throw new Error("Room type does not exists.");}
+			else {
+				// implement editing
+			}
+
+		} catch (error) {
+			throw error;
+		} finally {
+			this.#dbContext.close();
+		}
 	}
 
 	async deleteRoomType() {
@@ -111,10 +136,17 @@ export class RoomService {
 			if (!roomType) {
 				throw new Error("Room type does not exists.");}
 			else {
-				const result = await this.#dbContext.deleteValues(Tables.ROOMTYPES, { Id: this.#message.data.roomTypeId });
-				if (result.changes > 0) {
-					resObj.success = "Successfully deleted.";
-					return resObj;
+				//check the deletion of all rooms of the type
+				if(roomType.RoomsCount == 0){
+					const result = await this.#dbContext.deleteValues(Tables.ROOMTYPES, { Id: this.#message.data.roomTypeId });
+					if (result.changes > 0) {
+						resObj.success = "Successfully deleted.";
+						return resObj;
+				}
+				else{
+					throw new Error("All rooms of this type should be deleted.");
+				}
+				
 				} else {
 					throw new Error("An error occurred in deleting room type");
 				}
@@ -132,12 +164,34 @@ export class RoomService {
 
 		let resObj = {};
 		const data = this.#message.data;
+		let rTypeObj = {};
+		let rTypeList = [];
 		//if user is an hotel owner
 		try {
 			this.#dbContext.open();
 			let rows = await this.#dbContext.getValues(Tables.ROOMTYPES, {HotelId: data.HotelId});
+			for (const rType of rows) {
+				rTypeObj.Code = rType.Code;
+				rTypeObj.RoomCount = rType.RoomCount;
+				rTypeObj.SqftSqft = rType.Sqft;
+				rTypeObj.Price = rType.Price;
+				rTypeObj.Facilities = rType.Facilities;
+				 let sleeps = 0;
+				 if(rType.SingleBedCount>0){
+					sleeps += SingleBedCount*1
+				 }
+				 if(rType.DoubleBedCount>0){
+					sleeps += DoubleBedCount*2
+				 }
+				 if(rType.TripleBedCount>0){
+					sleeps += TripleBedCount*3
+				 }
+				rTypeList.Sleeps = sleeps
 
-			resObj.success = rows;
+				rTypeList.push(rTypeObj)
+
+			}
+			resObj.success = rTypeList;
 			return resObj;
 		} catch (error) {
 			throw error;
